+++ demosplan/DemosPlanCoreBundle/DataGenerator/Factory/Workflow/PlaceFactory.php
@@ -13,6 +13,7 @@ namespace demosplan\DemosPlanCoreBundle\DataGenerator\Factory\Workflow;
 use demosplan\DemosPlanCoreBundle\DataGenerator\Factory\Procedure\ProcedureFactory;
 use demosplan\DemosPlanCoreBundle\Entity\Workflow\Place;
 use demosplan\DemosPlanCoreBundle\Repository\Workflow\PlaceRepository;
+use Doctrine\ORM\EntityRepository;
 use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Persistence\Proxy;
 use Zenstruck\Foundry\Persistence\ProxyRepositoryDecorator;
@@ -35,6 +36,22 @@ use Zenstruck\Foundry\Persistence\ProxyRepositoryDecorator;
  * @method static Place[]|Proxy[]                          findBy(array $attributes)
  * @method static Place[]|Proxy[]                          randomRange(int $min, int $max, array $attributes = [])
  * @method static Place[]|Proxy[]                          randomSet(int $number, array $attributes = [])
+ *
+ * @phpstan-method        Place&Proxy<Place> create(array|callable $attributes = [])
+ * @phpstan-method static Place&Proxy<Place> createOne(array $attributes = [])
+ * @phpstan-method static Place&Proxy<Place> find(object|array|mixed $criteria)
+ * @phpstan-method static Place&Proxy<Place> findOrCreate(array $attributes)
+ * @phpstan-method static Place&Proxy<Place> first(string $sortedField = 'id')
+ * @phpstan-method static Place&Proxy<Place> last(string $sortedField = 'id')
+ * @phpstan-method static Place&Proxy<Place> random(array $attributes = [])
+ * @phpstan-method static Place&Proxy<Place> randomOrCreate(array $attributes = [])
+ * @phpstan-method static ProxyRepositoryDecorator<Place, EntityRepository> repository()
+ * @phpstan-method static list<Place&Proxy<Place>> all()
+ * @phpstan-method static list<Place&Proxy<Place>> createMany(int $number, array|callable $attributes = [])
+ * @phpstan-method static list<Place&Proxy<Place>> createSequence(iterable|callable $sequence)
+ * @phpstan-method static list<Place&Proxy<Place>> findBy(array $attributes)
+ * @phpstan-method static list<Place&Proxy<Place>> randomRange(int $min, int $max, array $attributes = [])
+ * @phpstan-method static list<Place&Proxy<Place>> randomSet(int $number, array $attributes = [])
  */
 final class PlaceFactory extends PersistentProxyObjectFactory
 {
@@ -43,6 +60,9 @@ final class PlaceFactory extends PersistentProxyObjectFactory
         return Place::class;
     }
 
+    /**
+     * @see https://symfony.com/bundles/ZenstruckFoundryBundle/current/index.html#model-factories
+     */
     protected function defaults(): array|callable
     {
         return [
@@ -54,6 +74,9 @@ final class PlaceFactory extends PersistentProxyObjectFactory
         ];
     }
 
+    /**
+     * @see https://symfony.com/bundles/ZenstruckFoundryBundle/current/index.html#initialization
+     */
     protected function initialize(): static
     {
         return $this;