+++ demosplan/DemosPlanCoreBundle/ResourceTypes/OriginalStatementResourceType.php
@@ -21,13 +21,17 @@ use demosplan\DemosPlanCoreBundle\Entity\Statement\Statement;
 use demosplan\DemosPlanCoreBundle\Event\IsOriginalStatementAvailableEvent;
 use demosplan\DemosPlanCoreBundle\Exception\UndefinedPhaseException;
 use demosplan\DemosPlanCoreBundle\Logic\ApiRequest\ResourceType\DplanResourceType;
+use demosplan\DemosPlanCoreBundle\Logic\ApiRequest\ResourceType\ReadableEsResourceTypeInterface;
 use demosplan\DemosPlanCoreBundle\Logic\FileService;
 use demosplan\DemosPlanCoreBundle\Logic\Statement\StatementProcedurePhaseResolver;
 use demosplan\DemosPlanCoreBundle\Logic\Statement\StatementService;
 use demosplan\DemosPlanCoreBundle\Repository\FileContainerRepository;
 use demosplan\DemosPlanCoreBundle\ResourceConfigBuilder\OriginalStatementResourceConfigBuilder;
+use demosplan\DemosPlanCoreBundle\Services\Elasticsearch\AbstractQuery;
+use demosplan\DemosPlanCoreBundle\Services\Elasticsearch\QueryStatement;
 use EDT\JsonApi\ResourceConfig\Builder\ResourceConfigBuilderInterface;
 use EDT\PathBuilding\End;
+use Elastica\Index;
 
 /**
  * @template-extends DplanResourceType<StatementInterface>
@@ -39,9 +43,10 @@ use EDT\PathBuilding\End;
  * @property-read StatementResourceType $movedStatement
  * @property-read StatementResourceType $parentStatementOfSegment Do not expose! Alias usage only.
  */
-final class OriginalStatementResourceType extends DplanResourceType implements OriginalStatementResourceTypeInterface
+final class OriginalStatementResourceType extends DplanResourceType implements OriginalStatementResourceTypeInterface, ReadableEsResourceTypeInterface
 {
     public function __construct(
+        private readonly QueryStatement $esQuery,
         private readonly FileService $fileService,
         private readonly StatementService $statementService,
         private readonly StatementProcedurePhaseResolver $statementProcedurePhaseResolver,
@@ -155,11 +160,36 @@ final class OriginalStatementResourceType extends DplanResourceType implements O
                 return $fileContainers;
             });
 
+        $originalStatementConfig->procedure
+            ->setRelationshipType($this->resourceTypeStore->getProcedureResourceType())
+            ->setReadableByPath()
+            ->setFilterable();
+
         return $originalStatementConfig;
     }
 
     private function hasAccessPermissions(): bool
     {
         return $this->currentUser->hasPermission('feature_json_api_original_statement');
+    }
+
+    public function getQuery(): AbstractQuery
+    {
+        return $this->esQuery;
+    }
+
+    public function getScopes(): array
+    {
+        return $this->esQuery->getScopes();
+    }
+
+    public function getSearchType(): Index
+    {
+        return $this->jsonApiEsService->getElasticaTypeForTypeName(self::getName());
+    }
+
+    public function getFacetDefinitions(): array
+    {
+        return [];
     }
 }