+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add Versioning of custom fields of segments
 
 ## UNRELEASED
+- Export Original Statements as csv in the Statement List
 ## v4.1.0 (2025-05-21)
 ## v4.0.0 (2025-05-21)
 - Update to symfony 6.4